[UPD] Start of OMS class refactoring
